Fixed upcoming ICOs
